cable refacto wip - basics working!
* fully remove PortIdentity
* stop passing/storing ids, use pointers directly
* stop trying to orchestrate everything via the gamemode
* much better connect/disconnect/persist/destroy handling

the cable actor was previously a pretty simple thing, just meant to draw
a debug line and some circles between two points. although not entirely
intentional from the start of this refactor, the cable actor is now the
first-class citizen it should be (along with modules). the chain of
control now starts at the avatar/controller asking the game mode for a
cable, and then telling the cable which ports to interact with. the
cable is then mostly responsible for connects, disconnects, and
requesting persisting/depersisting. finally, the avatar/controller
handling the request for destruction of the cable.