working threaded svg render (post testing)
the initial solution was attempting to allocate the unsigned char* rgba
array in the thread, then delete it from outside of the thread. this
leads to weird problems. instead, the renderer creates a texture transient
and passes it off to the threaded worker. the worker then handles updating
the transient and setting a boolean when it is done. the renderer watches
for this boolean change and passes the new texture off for long-term
storage.